// src/types/common/Pokemon.ts
export interface PokemonProps {
  id: number;
  num: string;
  name: string;
  img: string;
  type: string[];
  height: string;
  weight: string;
  candy: string;
  candy_count: number;
  egg: string;
  spawn_chance: number;
  avg_spawns: number;
  spawn_time: string;
  multipliers: number[];
  weaknesses: string[];
  next_evolution?: SimplifiedPokemonProps[];
  prev_evolution?: SimplifiedPokemonProps[];
}

interface SimplifiedPokemonProps {
  name: string;
  num: string;
}